t/t-fsck.sh: add files in directory to object test
In our "fsck detects invalid objects" test of the "git lfs fsck"
command we currently only use files in the top-level directory
for which we then corrupt or remove the corresponding Git LFS
object files.

In a subsequent PR we will add a parallel test which checks that
the "lfs.fetchexclude" option is supported for this command,
specifically by excluding files in a subdirectory.  For the
primary test, then, we want to demonstrate normal behaviour
without path-based exclusions, so we add files in a subdirectory
to our existing test and ensure their corruption is detected.