lfs,t: use prune exclude filter on previous SHAs
In commit d2221dcecacc6a2ad38ffd2e429fca18805cb4ea of PR #2851
the "git lfs prune" command was changed to respect the
"lfs.fetchexclude" configuration option such that objects would
always be pruned if they were referenced by files whose paths
matched one of the patterns in the configuration option (unless
they were referenced by an unpushed commit).

However, while this filter was applied to files referenced by
recent refs (including HEAD), it was not applied to files referenced
only by recent commits previous to a recent ref.  This can result in
the unexpected consequence that an object in HEAD is pruned because
it is referenced only by a file that matches the "lfs.fetchexclude"
filter, but a recent previous version of the object is not pruned
despite also only being referenced by the same file.

We therefore add equivalent filtering to the logPreviousSHAs()
internal function which is called by the ScanPreviousVersions()
GitScanner method used by pruneTaskGetPreviousVersionsOfRef()
in the "git lfs prune" command's main phase.

(Note that in PR #1743 the Git log scanning functions were refactored,
and a common parseScannerLogOutput() internal function was created to
be used when performing log scans during "git lfs prune" commands.
This replaced the original logPreviousSHAs() function, which had
support for path-based filtering; however, that functionality was
apparently never used.)

We also add a test which confirms that "git lfs prune" respects
the "lfs.fetchexclude" configuration option insofar as it prunes
Git LFS objects for files whose paths match a pattern in the filter,
both when they appear in commits directly referenced by a recent ref
and when they only appear in commits previous to those.

Note too that we add one extra check to the existing "prune
unreferenced and old" test for consistency with our new test.