subprocess: add function to replace percent codes
In the Git merge driver code, Git uses percent sequences (à la
printf(3)) to handle the file names passed to the merge driver.  We'll
want to do the same thing when allowing the user to perform a merge on
their own using a custom program, so let's add a function that replaces
percent sequences in a string with their specified replacement,
shell-quoting each argument along the way.  This uses the same algorithm
that Git does.