t/t-prune.sh: add tag check to unpushed test
In PR #742 the "git lfs prune" command was introduced along with
accompanying tests, one of which is the "prune keep unpushed"
test that checks whether Git LFS objects referenced by not-yet-
pushed commits are always retained by "git lfs prune".

In commit 978899e3b538bdd7df0873c2b50d229d5748605d of that PR
the initial version of the "prune" tests included some fixture
data with commits that were referenced via a tag only, not via
a branch ref.  However, the test never confirmed that the Git LFS
objects in these commits were retained by the "prune" command.

This initial test was then refactored into several tests in
commits 03b85e08acf727428c4e9c0d805f5a8a3da501e7 and
58dfa23c9e87bd638eb2bd7b4c8e208eab8031e7, including the "prune
keep unpushed" one, and in the process two lines were left
which referred to the fixture data ("oid_keepunpushedtagged1"
and "oid_keepunpushedtagged2") but this data was otherwise
dropped from the test, and no tag check was implemented.

We therefore re-introduce some fixture data for this test
which simulates a tag that points to a commit from a deleted
branch, so that the tag is the only reference to this commit
and its ancestors.  We then ensure that "git lfs prune" retains
all the Git LFS objects from these commits, even if they
are not recent, when they have not been pushed.  Once they are
pushed, we then confirm that only the object in the tagged commit
is retained (because it is a referenced by a recent ref).

We also update a comment in the "prune unreferenced and old"
test which refers to a filename in its fixture data that
corresponds to how it was originally defined in commit
978899e3b538bdd7df0873c2b50d229d5748605d, but which was
subsequently changed (at the same time the comment was added)
in commit 03b85e08acf727428c4e9c0d805f5a8a3da501e7.