update to Go 1.18 and drop older version support
As our official policy is to support the latest version of Go, we
upgrade our CI and release workflows to version 1.18, which was
recently released.

We also drop support for versions older than 1.17 as they are
not supported by upstream Go, and because the mitigations for
CVE-2022-27191 are only available in Go 1.17 and later.

Specifically, the updates to the x/crypto/ssh package (which Git LFS
does not use, so this security issue does not strictly affect Git LFS)
from version v0.0.0-20220314234659-1baeb1ce4c0b depend on changes in
the x/net package which are not available in any versions of that
package prior to v0.0.0-20211112202133-69e39bad7dc2.  See for reference:

https://groups.google.com/g/golang-announce/c/-cp44ypCT5s