commands,lfshttp,ssh,t: improve SSH URL parsing
Git LFS supports both "bare" SSH URLs, of the form "user@host:path",
and SSH URIs of the form "ssh://user@host/path", when parsing its
configuration fields.  However, we do not at present retain any
record of this difference, and so we are not able to reconstruct
the original input when reporting SSH endpoints in the "git lfs env"
command.

We also do not report a non-default SSH port, if one was configured,
in the output from that command, leading to confusion when the
input configuration was a URI like "ssh://example.com:1337/path"
and we report it as "example.com:/path", which looks like we have
dropped the port entirely.

We therefore add a URIFormat member to our SSHMetadata structure
and set this true only when parsing an SSH URI with a "ssh://"
scheme.  We also add a URLFromMetadata() helper function in our
"ssh" package which the "git lfs env" command can now use to
format SSH metadata to match its original input format, including
any non-default port number.

As well, we revise one change which was introduced in commit
a0190a60205bc6223d1d9028c34585583721c of PR #4446, whereby the
lfshttp.EndpointFromBareSshUrl() function always strips off the
leading path separator.  The code actually only performs this
action if the path returned by EndpointFromSshUrl() begins with
a slash character, but since that function expects an SSH URI
and we construct one as its argument, it will always return an
absolute path with a leading separator, and so we always strip
that off.

Instead, we now detect whether the configured path in the bare
SSH URL is an absolute or relative one before we generate a
temporary URI to pass to EndpointFromSshUrl().  We can then remove
the leading slash always returned by that function if the original
input had a relative path and not otherwise.

We also expand an existing test in the t/t-env.sh file so as to
confirm that our changes work as expected with a variety of
bare SSH URLs.