Makefile: use pax format for tar archives
There are a wide variety of possible formats for tar archives.  The
POSIX standard provides the ustar format (Unix standard tar), and GNU
tar has extended this with various extensions to better support long
file names.  The latter is the default format for GNU tar.

However, POSIX has also standardized an additional format, called the
pax format, which is an extension to the ustar format that allows for
functionally unlimited file name lengths, file sizes, and other
attributes.  Because it is standardized, it is substantially more
portable than GNU tar, and it is also used by default in Git, so anyone
who can consume a Git-produced tar archive can therefore handle this
format.

The pax format can be enabled with the --posix option to both GNU and
BSD tar.  Since it's a substantial improvement over the default for GNU
tar, let's produce it by default when we create a tar archive.