ls-files: add a --json option
We'd like to have a machine-readable option for ls-files in order to
avoid ambiguity in parsing.  Let's add a --json option to provide output
in JSON format.

Note that we use a JSON object with a single item here in order to allow
us to emit more data in the future without breaking backward
compatibility.