docs/man: fix titles of AsciiDoc man pages
When pandoc did our conversion to AsciiDoc, it was not aware that we
were working with manual pages.  Automatic handling of manual pages with
Asciidoctor requires that the title of the manual page be a top-level,
not a second-level, title, and that the name section be explicitly
specified.

Let's fix our manual pages to meet that standard automatically with the
following shell command.  This has the pleasant side effect of also
normalizing whether we use one or two dashes between the name and the
description to a single dash, which is what the Git manual pages use.

  ruby -pi -e '$_.gsub!(/^(=+)=( \S)/, "\\1\\2")' \
  -e '$_.gsub!(/^= (\S+)(\(\d\)) -+ (.*)$/, "= \\1\\2\n\n== NAME\n\n\\1 - \\3")' \
  *.adoc