installer: handle `BashOnly` Git for Windows gracefully
It was reported in https://github.com/git-lfs/git-lfs/discussions/5031
that Git LFS' installer aborts because it does not find the `git`
executable when Git for Windows was installed with the "Bash Only" PATH
option, i.e. in the mode where `PATH` is not modified at all.

Detect this situation, and fall back to implicitly extend the `PATH` to
find Git for Windows' `git.exe`.

Co-authored-by: Chris Darroch <chrisd@apache.org>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>