doc: fix excluded parameters in options
When we autoconverted the ronn source to AsciiDoc, we used the
CommonMark converter, since there isn't a ronn-specific one.  However,
that caused some problems because anything in angle brackets was assumed
to be an HTML tag and removed.

Fix all of the options descriptions that had a missing argument by
restoring the argument to what it was before.