rpm: set VERSION_ID to major version
The version ID set right now is a value like "8.5" on Rocky Linux 8.  In
such a situation, it's hard to compare, and this prevents rubygems-devel
from being installed properly in such a situation.  To make things work
properly, strip off the decimal point and decimal value from the end of
the string.