tq: clear sshTransfer on deletion
In some cases, we can end up with a panic due to a pure SSH transfer
lacking any connections, but it's not clear what those cases are.  Let's
help improve a possible failure case by explicitly setting sshTransfer
to nil, since shutting down is the only case in which we explicitly set
the number of connections to 0.

This means that we'll get a better error message in such a case, since
we'll attempt a legacy SSH connection, which won't work, and that will
be reported to the user.