tools: don't match MINGW as Cygwin
The cygpath binary from Git for Windows has a bug: when the path
contains braces, apostrophes, or certain other special characters, it
strips them out when called from Go, but not from the command line.

This results in us having broken behaviour with those characters in the
current working directory.  To address this, no longer consider MINGW to
be a Cygwin-like environment.