ssh: fix multiplexing condition
In our SSH code, we check if the error is non-nil and then enable
multiplexing.  However, this is backwards: only if the error is nil do
we have a valid path and can we enable multiplexing.  Fix this condition
properly.

Since our tests will now often get additional arguments, update
lfs-ssh-echo to handle them gracefully.