commands/uploader.go: update object skip comments
In commit 338ab406ba35a0645bc660f8cb24dad8913f9490 of PR #1812 the
prepareUpload() function was refactored so that it no longer checked
whether an object to be pushed had a local file copy or not and skipped it
if it did not.

(That functionality already existed in the uploadTransfer() function
in that it called the ensureFile() function before returning a
tq.Transfer object.  Note that these functions were in the
commands/commands.go source file at the time.  The uploadTransfer()
function is called for each pointer blob object by uploadPointers().)

However, commit 338ab406ba35a0645bc660f8cb24dad8913f9490 did not
update the comment in prepareUpload() which stated that the function
skipped any objects which did not have local file copies, so we now
revise that comment and add one in the uploadTransfer() function.