Add the FAQ to the documentation
We generally get a lot of requests for problems that are already
answered in the FAQ.  Right now, this lives in the wiki, but this is
hard to find, so let's include it in the documentation with Git LFS
itself.

Add the relevant portions to the binary and Makefile so that it can be
accessed using `git lfs help faq`, and add it as well to the Linux
packages.