docs: accurately describe the install --skip-repo option
The name of the `--skip-repo` option leads people to believe that it's
incompatible with `--local`.  Let's describe the option more accurately
so that people know exactly what it does.  Don't specifically mention
the global LFS filters because one can indeed use local LFS filters with
it as well.  Capitalize "LFS", and specifically mention that any of the
destination options may be used in conjunction with this option.