Use --sparse parameter for ls-files for performance optimization
When lfs is used with a huge mono repository with sparse-checkout,
the lfs post-commit command took a long time. Per default
git ls-files, which excute during the post-commit, retrieves all
file paths, irrespectevly of the sparse set. When sparse-index is
used ls-files respects the --sparse parameter, which omits the
irrelevant files that are not in the sparse set, resulting in a
performance improvement.