rpm: drop Ruby 2.x install and spec file
In commit git-lfs/build-dockers@544e07ee1a2d70eb85d1682202aac2906661735e
of PR git-lfs/build-dockers#47 we updated our CentOS 7 Dockerfile to
install Ruby 3.0 using the rh-ruby30-{ruby,build} packages and then run
our CentOS build script with that version of Ruby as the default.  We made
this change to support our use of Asciidoctor in preference to ronn,
which was completed in PR #5054, and as part of that PR, in commit
db9a82132a2bb066876d8ddf06c5255da2f199a4 we replaced our installation of
ronn with one of asciidoctor in our rpm/build_rpms.bsh script.

As a result, none of our CentOS or Rocky Linux container image builds
now rely on Ruby 2.x, so we can remove the code which installed that
version.

We can also remove the related code which used to install Ruby 2.x when
an even older version was the default, along with our local ruby.spec
file for Ruby 2.2.2.

Finally, we can remove the code which installed the rubygems-devel
package on CentOS 8 because that is taken care of by the corresponding
Dockerfile in our git-lfs/build-dockers project, which, like the
one for Rocky Linux 9, installs ruby, ruby-devel, and rubygems-devel.