rpm/build_rpms.bsh: drop CentOS 5/6 Go install
Since the original introduction of the rpm/build_rpms.bsh script in
commit 4a7162777913489e7c70a3da3235dcd3eeec553b of PR #332 it has
checked for the absence of the "go" binary and, if it is not found,
installs the golang package, except on CentOS 5 where fetches and
patches a golang.spec file and then builds a local RPM using that
in order to install Go.

However, in commit b560b853a6e557195c0c68875218f45c89dce56b of PR #1298 and
in commit git-lfs/build-dockers@64a3a9fc4df3ca28920a4051435d55487832c5b7 of
PR git-lfs/build-dockers#3 we dropped support for CentOS 5.

And in commit 88430de792d96ec74846779aeb0530a4fb1ff49a of PR #654 we
updated our CentOS Dockerfiles (before they were moved to the
git-lfs/build-dockers repository) to download and install Go from the
googleapis.com service, and we still continue to do this but as of
commit git-lfs/build-dockers@c63850354edd262922a9f6196cdc8290533107d5
in PR git-lfs/build-dockers#35 we now use the golang.org service.

This direct download obviates the logic in our rpm/build_rpms.bsh
script to install the golang package, and the special handling for
CentOS 5 is also obsolete, so we remove this logic from our script.