rpm/build_rpms.bsh: drop unused NODEPS variable
In commit e2f770c8f5f2251976a83a5983c9ce170a764884 of PR #428
the rpm/build_rpms.bsh script was updated to run the rpmbuild command
with the --nodeps option when the NODEPS environment variable was defined
and set non-zero.

However, in commit 88430de792d96ec74846779aeb0530a4fb1ff49a of PR #654
the script was revised to always pass the --nodeps option to the
rpmbuild command, so the NODEPS environment variable no longer had
any effect, and so We can therefore remove our handling of this variable.