t-update: fix pre-push hook test
In some cases, we're echoing the old hook contents, but we're not
echoing it into a file, which means we aren't testing what we think we
are: the old hook contents.  Let's actually test that by echoing the
contents of the script to the hook file name so our test works
correctly.

While we're at it, we need to properly escape the dollar sign in `$@`,
since otherwise it gets substituted before we write the script.