hooks: mention core.hookspath instead of just .git/hooks
In some cases, users may use `core.hooksPath` to locate their hooks in a
different location.  However, right now, our hook and installation
instructions mention only `.git/hooks`.

We can update the instructions without a problem, but we don't want to
hard-code the hook path into the hooks themselves because of things like
symlinks and moved repository, plus the fact that we can't update a hook
automatically unless it's identical minus some whitespace changes.  To
avoid spuriously failing to update a hook, let's print the right
location with the instructions, and just mention `core.hookspath` and
`.git/hooks` in the message, leaving it to the user to discover.

Note that technically, core.hookspath was not implemented until Git
2.9.0.  However, while we still support older versions of Git, no
presently supported version of Ubuntu or Debian offers a version older
than 2.11, so the risk of practical confusion by mentioning this is low.
Nevertheless, make our tests pass by checking for an old Git explicitly.